[RENAME]: 일부 파일명 변경 && [STYLE]: 코드 순서 변경 작업
