+++ RGRG/Resources/UIcolor+CALayer.swift
@@ -9,23 +9,26 @@ import Foundation
 
 import UIKit
 
+// MARK: - UIColor
+
 extension UIColor {
-  convenience init(hex: String, alpha: CGFloat = 1.0) {
-    var hexSanitized = hex.trimmingCharacters(in: .whitespacesAndNewlines)
-    hexSanitized = hexSanitized.replacingOccurrences(of: "#", with: "")
+    convenience init(hex: String, alpha: CGFloat = 1.0) {
+        var hexSanitized = hex.trimmingCharacters(in: .whitespacesAndNewlines)
+        hexSanitized = hexSanitized.replacingOccurrences(of: "#", with: "")
 
-    var rgb: UInt64 = 0
+        var rgb: UInt64 = 0
 
-    Scanner(string: hexSanitized).scanHexInt64(&rgb)
+        Scanner(string: hexSanitized).scanHexInt64(&rgb)
 
-    let red = CGFloat((rgb & 0xFF0000) >> 16) / 255.0
-    let green = CGFloat((rgb & 0x00FF00) >> 8) / 255.0
-    let blue = CGFloat(rgb & 0x0000FF) / 255.0
+        let red = CGFloat((rgb & 0xFF0000) >> 16) / 255.0
+        let green = CGFloat((rgb & 0x00FF00) >> 8) / 255.0
+        let blue = CGFloat(rgb & 0x0000FF) / 255.0
 
-    self.init(red: red, green: green, blue: blue, alpha: alpha)
-  }
+        self.init(red: red, green: green, blue: blue, alpha: alpha)
+    }
 }
 
+// MARK: - CALayer
 
 extension CALayer {
     func addBorder(_ arr_edge: [UIRectEdge], color: UIColor, width: CGFloat) {
@@ -33,23 +36,18 @@ extension CALayer {
             let border = CALayer()
             switch edge {
             case UIRectEdge.top:
-                border.frame = CGRect.init(x: 0, y: 0, width: frame.width, height: width)
-                break
+                border.frame = CGRect(x: 0, y: 0, width: frame.width, height: width)
             case UIRectEdge.bottom:
-                border.frame = CGRect.init(x: 0, y: frame.height - width, width: frame.width, height: width)
-                break
+                border.frame = CGRect(x: 0, y: frame.height - width, width: frame.width, height: width)
             case UIRectEdge.left:
-                border.frame = CGRect.init(x: 0, y: 0, width: width, height: frame.height)
-                break
+                border.frame = CGRect(x: 0, y: 0, width: width, height: frame.height)
             case UIRectEdge.right:
-                border.frame = CGRect.init(x: frame.width - width, y: 0, width: width, height: frame.height)
-                break
+                border.frame = CGRect(x: frame.width - width, y: 0, width: width, height: frame.height)
             default:
                 break
             }
-            border.backgroundColor = color.cgColor;
+            border.backgroundColor = color.cgColor
             self.addSublayer(border)
         }
     }
 }
-
