--- RGRG/Scenes/Recuitment/RecuitmentViewController.swift
@@ -1,8 +1,0 @@
-//
-//  MatePageViewController.swift
-//  RGRG
-//
-//  Created by (^ㅗ^)7 iMac on 2023/10/12.
-//
-
-import Foundation
